Verlate additie
+random method
#vormen parameter veranderd naar random
+Extra paneel toegevoegd voor score
+Extra paneel toegevoegd in hoofdpaneel

+++ Eindopdracht 2/src/reactie_game/Biljartbal.java
@@ -8,7 +8,7 @@ public class Biljartbal extends Circle {
 	private final double maxX, maxY;
 	private double dx, dy;
 
-	public Biljartbal(int x, int y, int straal, Color kleur, double maxX, double maxY) {
+	public Biljartbal(double x, double y, int straal, Color kleur, double maxX, double maxY) {
 		super(x, y, straal);
 		this.straal = straal;
 		this.maxX = maxX;
